Split transaction status polling out of TransactionsRepositoryImpl
Extract TransactionStateService (single poll step: status fetch, state
policy, hash-change handling) and TransactionStateScheduler (poll loop,
job dedup, backoff) from the repository, mirroring the iOS structure.
The repository is plain Room CRUD again.

Replace the changedTransactions StateFlow bus and GetChangedTransactions
with direct TransactionPostProcessingService calls, so completed
transactions cannot be conflated or replayed to late subscribers.

Wire the scheduler into AppLifecycleCoordinator: polling stops in the
background and resumes from the initial interval on foreground.
