+++ android/app/src/main/kotlin/com/gemwallet/android/AppLifecycleCoordinator.kt
@@ -6,6 +6,7 @@ import com.gemwallet.android.data.repositories.connection.ConnectionStatusObserv
 import com.gemwallet.android.data.repositories.device.DeviceObserverService
 import com.gemwallet.android.data.repositories.perpetual.HyperliquidObserverService
 import com.gemwallet.android.data.repositories.stream.StreamObserverService
+import com.gemwallet.android.data.repositories.transactions.TransactionStateScheduler
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -16,6 +17,7 @@ class AppLifecycleCoordinator @Inject constructor(
     private val hyperliquidObserver: HyperliquidObserverService,
     private val nodeAuthTokenService: NodeAuthTokenService,
     private val connectionStatusObserver: ConnectionStatusObserver,
+    private val transactionStateScheduler: TransactionStateScheduler,
 ) : DefaultLifecycleObserver {
 
     override fun onStart(owner: LifecycleOwner) {
@@ -24,6 +26,7 @@ class AppLifecycleCoordinator @Inject constructor(
         hyperliquidObserver.start()
         nodeAuthTokenService.start()
         connectionStatusObserver.start()
+        transactionStateScheduler.start()
     }
 
     override fun onStop(owner: LifecycleOwner) {
@@ -32,5 +35,6 @@ class AppLifecycleCoordinator @Inject constructor(
         hyperliquidObserver.stop()
         nodeAuthTokenService.stop()
         connectionStatusObserver.stop()
+        transactionStateScheduler.stop()
     }
 }
