+++ android/data/repositories/src/main/kotlin/com/gemwallet/android/data/repositories/assets/TransactionPostProcessingService.kt
@@ -1,6 +1,5 @@
 package com.gemwallet.android.data.repositories.assets
 
-import com.gemwallet.android.application.transactions.coordinators.GetChangedTransactions
 import com.gemwallet.android.cases.nft.SyncNfts
 import com.gemwallet.android.cases.stake.SyncStakeDelegations
 import com.gemwallet.android.ext.getAssociatedAssetIds
@@ -10,40 +9,20 @@ import com.gemwallet.android.model.TransactionExtended
 import com.wallet.core.primitives.Transaction
 import com.wallet.core.primitives.TransactionType
 import com.wallet.core.primitives.WalletId
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class TransactionPostProcessingService(
-    getChangedTransactions: GetChangedTransactions,
+class TransactionPostProcessingService @Inject constructor(
     private val assetsRepository: AssetsRepository,
     private val syncStakeDelegations: SyncStakeDelegations,
     private val syncNfts: SyncNfts,
-    private val scope: CoroutineScope,
 ) {
-
-    @Inject
-    constructor(
-        getChangedTransactions: GetChangedTransactions,
-        assetsRepository: AssetsRepository,
-        syncStakeDelegations: SyncStakeDelegations,
-        syncNfts: SyncNfts,
-    ) : this(getChangedTransactions, assetsRepository, syncStakeDelegations, syncNfts, CoroutineScope(Dispatchers.IO))
-
-    init {
-        scope.launch(Dispatchers.IO) {
-            getChangedTransactions.getChangedTransactions().collect {
-                processTransactions(it)
-            }
-        }
-    }
 
     internal suspend fun processTransactions(transactions: List<TransactionExtended>) = withContext(Dispatchers.IO) {
         transactions.map { transactionExtended ->
