--- android/gemcore/src/main/kotlin/com/gemwallet/android/application/transactions/coordinators/GetChangedTransactions.kt
@@ -1,9 +1,0 @@
-package com.gemwallet.android.application.transactions.coordinators
-
-import com.gemwallet.android.model.TransactionExtended
-import kotlinx.coroutines.flow.Flow
-
-interface GetChangedTransactions {
-
-    fun getChangedTransactions(): Flow<List<TransactionExtended>>
-}
